refactor(purchase-order): HqPurchaseOrderCreateView 비대 SFC 책임 단위 분리 1차
1625줄 단일 SFC 를 가이드 위반(1000줄 금지) 정정 차원에서 분리.
재사용 가능한 4개 composable + 1개 아이콘 모듈로 추출, 부모 view 의
역할은 layout + 상태 조립으로 좁힘. 부모 1625 → 1482 (-143).

- icons.js: 인라인 7개 SVG 정의 모듈화
- useToast: 토스트 + 자동 cleanup
- useFacets: 옵션 facet 필터 상태/로직
- usePurchaseOrderStockSim: SKU 단위 재고 시뮬레이션
- usePurchaseOrderDraft: localStorage 드래프트 영속화

카탈로그/장바구니/모달 추출은 후속 사이클.
