+++ src/components/hq/purchase-order/icons.js
@@ -1,0 +1,59 @@
+import { h } from 'vue'
+
+const IconBase = (paths) => ({
+  props: {
+    size: { type: Number, default: 16 },
+    strokeWidth: { type: Number, default: 2 },
+  },
+  render() {
+    return h(
+      'svg',
+      {
+        width: this.size,
+        height: this.size,
+        viewBox: '0 0 24 24',
+        fill: 'none',
+        stroke: 'currentColor',
+        'stroke-width': this.strokeWidth,
+        'stroke-linecap': 'round',
+        'stroke-linejoin': 'round',
+        'aria-hidden': 'true',
+      },
+      paths.map((p) => h(p.tag, p.attrs)),
+    )
+  },
+})
+
+export const SearchIcon = IconBase([
+  { tag: 'circle', attrs: { cx: '11', cy: '11', r: '7' } },
+  { tag: 'path', attrs: { d: 'm20 20-3.5-3.5' } },
+])
+
+export const PlusIcon = IconBase([
+  { tag: 'path', attrs: { d: 'M12 5v14' } },
+  { tag: 'path', attrs: { d: 'M5 12h14' } },
+])
+
+export const TrashIcon = IconBase([
+  { tag: 'polyline', attrs: { points: '3 6 5 6 21 6' } },
+  { tag: 'path', attrs: { d: 'M19 6 18 20a2 2 0 0 1-2 2H8a2 2 0 0 1-2-2L5 6' } },
+  { tag: 'path', attrs: { d: 'M10 11v6' } },
+  { tag: 'path', attrs: { d: 'M14 11v6' } },
+])
+
+export const ArrowLeftIcon = IconBase([
+  { tag: 'path', attrs: { d: 'M19 12H5' } },
+  { tag: 'path', attrs: { d: 'm12 19-7-7 7-7' } },
+])
+
+export const ShoppingCartIcon = IconBase([
+  { tag: 'circle', attrs: { cx: '9', cy: '21', r: '1' } },
+  { tag: 'circle', attrs: { cx: '20', cy: '21', r: '1' } },
+  { tag: 'path', attrs: { d: 'M1 1h4l2.7 13.4a2 2 0 0 0 2 1.6h9.7a2 2 0 0 0 2-1.6L23 6H6' } },
+])
+
+export const AlertTriangleIcon = IconBase([
+  { tag: 'path', attrs: { d: 'M10.29 3.86 1.82 18a2 2 0 0 0 1.71 3h16.94a2 2 0 0 0 1.71-3L13.71 3.86a2 2 0 0 0-3.42 0Z' } },
+  { tag: 'path', attrs: { d: 'M12 9v4' } },
+  { tag: 'path', attrs: { d: 'M12 17h.01' } },
+])
